refactor: extract base avatar logic into mixin
